Fixed integer overflow issue in problem 3.

+++ 003/prime-factors.c
@@ -3,8 +3,8 @@
 
 // find the largest prime factor of 600851475143
 
-bool isPrime(unsigned long int i) {
-	unsigned long int n;
+bool isPrime(unsigned long long int i) {
+	unsigned long long int n;
 	for (n = 2; n <= i/2; n++) {
 		if (i%n == 0)
 			return false;
@@ -12,16 +12,16 @@ bool isPrime(unsigned long int i) {
 	return true;
 }
 int main(int argc, const char *argv[]) {
-	unsigned long int n = 600851475143;
-	unsigned long int i = 2;
+	unsigned long long int n = 600851475143ULL;
+	unsigned long long int i = 2;
 	while(!isPrime(n)) {
 		if (isPrime(i) && n%i == 0) {
 			n /= i;
-			printf("%ld\n", i);
+			printf("%lld\n", i);
 		} else {
 			i++;
 		}
 	}
-	printf("%ld\n", n);
+	printf("%lld\n", n);
 	return 0;
 }
